# api/swagger/swagger.yaml
swagger: '2.0'
info:
  version: '1.0'
  title: Patient Data REST API
  description: >-
    Backend REST API for WeCare application to manage patients and patients'
    record
  contact: {
    
  }
host: wecare-api-supriya-raj.herokuapp.com
basePath: '/'
securityDefinitions: {}
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
  /patients:
    get:
      description: Fetch all patients from database
      summary: Get All Patients
      tags:
        - Patients
      operationId: GetAllPatients
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          required: false
          default: 'Bearer {token}'
          type: string
      responses:
        '200':
          description: ''
          headers: {}
    post:
      description: Add new patient to database
      summary: Create New Patient
      tags:
        - Patients
      operationId: CreateNewPatient
      deprecated: false
      produces:
        - application/json
      consumes:
        - application/json
      parameters:
        - name: Authorization
          in: header
          required: false
          default: 'Bearer {token}'
          type: string
        - name: Content-Type
          in: header
          required: true
          type: string
          description: JSON Type
        - name: Body
          in: body
          required: true
          description: ''
          schema:
            $ref: '#/definitions/CreateNewPatientRequest'
      responses:
        '200':
          description: ''
          headers: {}
  /patients/5d725a037b292f5f8ceff707:
    get:
      description: Get single patient by id
      summary: Get Single Patient
      tags:
        - Patients
      operationId: GetSinglePatient
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          required: false
          default: 'Bearer {token}'
          type: string
      responses:
        '200':
          description: ''
          headers: {}
    put:
      description: Updates patient and database
      summary: Update Patient
      tags:
        - Patients
      operationId: UpdatePatient
      deprecated: false
      produces:
        - application/json
      consumes:
        - application/json
      parameters:
        - name: Authorization
          in: header
          required: false
          default: 'Bearer {token}'
          type: string
        - name: Content-Type
          in: header
          required: true
          type: string
          description: JSON Type
        - name: Body
          in: body
          required: true
          description: ''
          schema:
            $ref: '#/definitions/UpdatePatientRequest'
      responses:
        '200':
          description: ''
          headers: {}
  /patients/5fb6c6d047583b3b2f964b95:
    delete:
      description: Deletes patients from database
      summary: Delete Patient
      tags:
        - Patients
      operationId: DeletePatient
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          required: false
          default: 'Bearer {token}'
          type: string
      responses:
        '200':
          description: ''
          headers: {}
...
definitions:
  CreateNewPatientRequest:
    title: CreateNewPatientRequest
    example:
      name: Test74 name
      age: 40
      phone: (555) 555-5555
      email: test74@gmail.com
      address: 350 Parliament Street Toronto Ontario M4Y 1GX
      diagnosis: E.coli infection
      description: >-
        Severe stomach cramps, diarrhea (often bloody), and vomiting.!!!! Lorem
        ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor
        incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam,
        quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo
        consequat.
    type: object
    properties:
      name:
        type: string
      age:
        type: integer
        format: int32
      phone:
        type: string
      email:
        type: string
      address:
        type: string
      diagnosis:
        type: string
      description:
        type: string
    required:
      - name
      - age
      - phone
      - email
      - address
      - diagnosis
      - description
  UpdatePatientRequest:
    title: UpdatePatientRequest
    example:
      address: 240 Wellesley Street East Toronto Ontario M4Y 1GX
    type: object
    properties:
      address:
        type: string
    required:
      - address
...